Fixed invisible gesture trail
It was using a shader that was not being compiled due to not being used in the scene or not being included in the Resources folder.
The material was being overwritten by a null value due to a the Start function coming after the UpdateRenderer. Made adjustments to prevent this.
Custom color and material selection for the gesture trail is enabled again.